added CategoryService, CategoryController; deleted CategoryValidator

+++ src/main/java/com/example/cosmocatsmarketplace/domain/Category.java
@@ -5,13 +5,13 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.UUID;
+
 
 @NoArgsConstructor
 @AllArgsConstructor
 @Data
 @Builder
 public class Category {
-    private int id;
+    private long id;
     private String name;
 }
